Move creation of DetailsFragment to newInstance() in companion object
